Backend/project_root/requirements_trash.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.0
cymem==2.0.10
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
nlpcloud==1.1.47
numpy==2.0.2
openai==1.57.4
packaging==24.2
pandoc==2.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
plumbum==1.9.0
ply==3.11
preshed==3.0.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.1.2
python-dotenv==1.0.1
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
setuptools==75.6.0
shellingham==1.5.4
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
thinc==8.3.2
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.0
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
